Move `IoUtils` from `x2cpg` to `io.shiftleft.utils`
